requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
pymongo
pymongo[srv]
certifi

#-e .